requirements.txt
# Code Review Crew - Requirements
# Python 3.8+

# Core Framework
crewai>=0.28.0
crewai-tools>=0.2.0

# LLM Providers (choose one or both)
openai>=1.0.0
anthropic>=0.18.0

# Configuration
pyyaml>=6.0.1

# Utipyyaml
crewai
python-dotenv
streamlit
gitpython>=1.0.0
rich>=13.0.0  # Better terminal output

# Optional: Enhanced features
# reportlab>=4.0.0  # For PDF generation (future)
# pylint>=3.0.0  # For static analysis (future)
# black>=23.0.0  # For code formatting checks (future)
